Language.Piet

Documentation

data PietError

Constructors

PietImageReaderError ImageReaderError

PietParserError ParserError

PietObjectGeneratorError ObjectGeneratorError

data PietStep

Constructors

StepReadImage

StepParse

StepMakeAssembly

StepGenerateExecutable

StepRunJIT

StepGenerateDOT

compile :: (MonadIO m, MonadError PietError m) => (PietStep -> m ()) -> ImageConfig -> OptimizationLevel -> FilePath -> FilePath -> m ()

Compile a Piet program.

run :: (MonadIO m, MonadError PietError m) => (PietStep -> m ()) -> ImageConfig -> OptimizationLevel -> FilePath -> m ()

Run a Piet program on JIT.

graphText :: (MonadIO m, MonadError PietError m) => (PietStep -> m ()) -> ImageConfig -> FilePath -> m Text

Convert a Piet program to a graph script.

nullReceiver :: Monad m => PietStep -> m ()

data OptimizationLevel

The optimization level to optimize assembly code.

Constructors

NoOptimization

O0

OptimizationLevelLow

O1

OptimizationLevelMiddle

O2

OptimizationLevelHigh

O3

SizeLevelLow

Os

SizeLevelHigh

Oz

data ImageReaderError

Constructors

ReadImageFileError String
The image file is unreadable.

UnsupportedImageError String
The input image has an unsupported format.

CodelSizeError
The specified size of codel is not fit for the image.

data AdditionalColorStrategy

This type is to determine how to deal with additional colors such as orange, gray, etc.

Constructors

AdditionalColorAsWhite
Treating as a white codel.

AdditionalColorAsBlack
Treating as a black codel.

AdditionalColorNearest
Treating as a codel which has the nearest color.

data MulticoloredCodelStrategy

This type is to determine how to deal with multicolored codels.

Constructors

MulticoloredCodelAsWhite
Treating as a white codel.

MulticoloredCodelAsBlack
Treating as a black codel.

MulticoloredCodelCenter
Picking up a center pixel.

MulticoloredCodelModal
Finding the modal color, the most frequent color.

MulticoloredCodelAverage
Calculating an average color.

data ImageConfig

Constructors

ImageConfig

Fields

additionalColor :: AdditionalColorStrategy

multicoloredCodel :: MulticoloredCodelStrategy

codelSize :: CodelSize

data CodelSize

Constructors

CodelSize Int

GuessCodelSize

data ObjectGeneratorError

Constructors

CompileError String
Failed to compile.

LinkError String
Failed to link object.

TempFileError String
Failed to create a temporary file.

data ParserError

Constructors

EmptyBlockTableError
The block table is empty.

IllegalInitialColorError
The initial codel of the block table is black.

MissingCodelIndexError Int
A codel index in the codel table is missing.