Language.Piet.ImageReader

Description

Functions to read images.

Documentation

data ImageReaderError

Constructors

ReadImageFileError String
The image file is unreadable.

UnsupportedImageError String
The input image has an unsupported format.

CodelSizeError
The specified size of codel is not fit for the image.

data AdditionalColorStrategy

This type is to determine how to deal with additional colors such as orange, gray, etc.

Constructors

AdditionalColorAsWhite
Treating as a white codel.

AdditionalColorAsBlack
Treating as a black codel.

AdditionalColorNearest
Treating as a codel which has the nearest color.

data MulticoloredCodelStrategy

This type is to determine how to deal with multicolored codels.

Constructors

MulticoloredCodelAsWhite
Treating as a white codel.

MulticoloredCodelAsBlack
Treating as a black codel.

MulticoloredCodelCenter
Picking up a center pixel.

MulticoloredCodelModal
Finding the modal color, the most frequent color.

MulticoloredCodelAverage
Calculating an average color.

data ImageConfig

Constructors

ImageConfig

Fields

additionalColor :: AdditionalColorStrategy

multicoloredCodel :: MulticoloredCodelStrategy

codelSize :: CodelSize

data CodelSize

Constructors

CodelSize Int

GuessCodelSize

readCodels :: (MonadIO m, MonadError ImageReaderError m) => ImageConfig -> FilePath -> m (Vector (Vector Codel))

Read an image and then convert to codels.

imageToCodels :: MonadError ImageReaderError m => ImageConfig -> DynamicImage -> m (Vector (Vector Codel))

Convert an image to codels.

rgbImageToCodels :: MonadError ImageReaderError m => ImageConfig -> Image PixelRGB8 -> m (Vector (Vector Codel))

Convert an image to codels. This function accepts only images in RGB8 format.