Language.Piet.Internal.Filler

Documentation

fillAll :: Eq a => Vector (Vector a) -> (Vector (Vector Int), IntMap [(Int, Int)])

Fill all connected pixels in a given image and return an filled image and a pixel map.

The filled image is consists of component indices. The pixel map contains coordinates of pixels which belong to a component whose index is the key.