Language.Piet.Parser

Description

Functions to parse images.

Documentation

data ParserError

Constructors

EmptyBlockTableError
The block table is empty.

IllegalInitialColorError
The initial codel of the block table is black.

MissingCodelIndexError Int
A codel index in the codel table is missing.

parse :: MonadError ParserError m => Vector (Vector Codel) -> m SyntaxGraph

Parse codels into a SyntaxGraph.

parseFilledImage :: MonadError ParserError m => (Vector (Vector (Codel, Int)), IntMap [(Int, Int)]) -> m SyntaxGraph

Parse a filled image which is returned by fillAll into a SyntaxGraph.