Language.Piet.Syntax

Description

A representation of the syntax of Piet.

Documentation

newtype Block

A representation of a codel block. This has information how to move into the next codel block.

Constructors

Block

Fields

nextBlockTable :: Map DPCC NextBlock

data SyntaxGraph

A representation of the syntax of Piet.

SyntaxGraph has a graph structure whose nodes represent codel blocks and edges represent the next steps.

Constructors

SyntaxGraph

Fields

_initialBlockIndex :: Int

_initialDPCC :: DPCC

_blockMap :: IntMap Block

EmptySyntaxGraph

data NextBlock

A data type which contains information about a move to the next block.

Constructors

NextBlock

Fields

_command :: Command

The command which will be executed when moving into the next codel block.

_dpcc :: DPCC

The DP and CC which will be used in the next codel block.

_blockIndex :: Int

The index of the next codel block.

ExitProgram

data DirectionPointer

Constructors

DPRight

DPDown

DPLeft

DPUp

data CodelChooser

Constructors

CCLeft

CCRight

data DPCC

Constructors

DPCC

Fields

getDP :: DirectionPointer

getCC :: CodelChooser

data Command

Constructors

NoOperation

Push Int

Pop

Add

Subtract

Multiply

Divide

Mod

Not

Greater

Pointer

Switch

Duplicate

Roll

InNumber

InChar

OutNumber

OutChar

commandFromTransition :: (Hue, Lightness) -> (Hue, Lightness) -> Int -> Command

showCommand :: Command -> String

showDPCC :: DPCC -> String